Refactor time tracking concerns in models
